update with suggested changes
