+++ src/layouts/page.tsx
@@ -95,26 +95,7 @@ export const contentfulPageQuery = graphql`
         }
       }
       entries {
-        ... on ContentfulTopBanner {
-          id
-          __typename
-          mainHeader
-          description
-          firstButtonText
-          firstButtonUrl
-          secondButtonText
-          secondButtonUrl
-          type
-          image {
-            localFile {
-              childImageSharp {
-                fluid(maxWidth: 1200, quality: 100) {
-                  ...GatsbyImageSharpFluid
-                }
-              }
-            }
-          }
-        }
+        ...ContentfulTopBannerFragment
         ...LargeTestimonialFragment
         ...FeatureGridFragment
         ...TestimonialCardsFragment
