+++ src/pages/index.tsx
@@ -144,6 +144,29 @@ export const LargeTestimonialFragment = graphql`
   }
 `;
 
+export const ContentfulTopBannerFragment = graphql`
+  fragment ContentfulTopBannerFragment on ContentfulTopBanner {
+    id
+    __typename
+    mainHeader
+    description
+    firstButtonText
+    firstButtonUrl
+    secondButtonText
+    secondButtonUrl
+    type
+    image {
+      localFile {
+        childImageSharp {
+          fluid(maxWidth: 1200, quality: 100) {
+            ...GatsbyImageSharpFluid
+          }
+        }
+      }
+    }
+  }
+`;
+
 export const TitleWithGraphicFragment = graphql`
   fragment TitleWithGraphicFragment on ContentfulTitleWithGraphic {
     id
@@ -321,26 +344,7 @@ export const indexPageQuery = graphql`
         node {
           title
           entries {
-            ... on ContentfulTopBanner {
-              id
-              __typename
-              mainHeader
-              description
-              firstButtonText
-              firstButtonUrl
-              secondButtonText
-              secondButtonUrl
-              type
-              image {
-                localFile {
-                  childImageSharp {
-                    fluid(maxWidth: 1200, quality: 100) {
-                      ...GatsbyImageSharpFluid
-                    }
-                  }
-                }
-              }
-            }
+            ...ContentfulTopBannerFragment
             ...FeatureGridFragment
             ...TestimonialCardsFragment
             ...ContentWithChecklistFragment
